back/requirements.txt
django==5.0.1
asgiref==3.7.2
sqlparse==0.4.4
psycopg2-binary==2.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
#Cors headers
django-cors-headers==4.3.1
# Image Handling
pillow==10.2.0
# Make request to 42 API
requests==2.31.0
#Monitoring
django-prometheus==2.3.1
python-logstash==0.4.8
#Websockets
daphne==4.1.0
channels==4.0.0
channels-redis==4.1.0
#Blockchain
web3==6.11.3
websockets==12.0
#One Time Password
pyotp==2.9.0
qrcode==7.4.2
#Pool Game
numpy==1.26.4
numpy-quaternion==2023.0.0